+++ components/wifi/wifi_lib_export/wifi.h
@@ -242,8 +242,14 @@ int     wifi_recv_ethernet_pkt_callback_reg(wifi_recv_ethernet_pkt_cb_t recv_cb)
 int     wifi_send_ethernet_pkt(uint8_t *data, int len, uint16_t retry_max, uint8_t retry_timeout);
 int     wifi_private_command(char *pvtcmd);
 
-void    wifi_rf_preprocess(void);
-void    wifi_rf_image_cal(void);
+/**
+ * wifi_rf_calibration
+ * 
+ * Merge function wifi_rf_preprocess() and function wifi_rf_image_cal(),
+ * and finally close RF!
+ * RF will be turned on after mac phy initialization is complete.
+*/
+void    wifi_rf_calibration(void);
 void    wifi_temp_cal_init(uint16_t adc_ch0_val,int8_t cap_comp);
 void    wifi_do_temp_cal_period(uint16_t adc_ch0_val);